refactor: make gallery content component reusable
