+++ src/components/base-gallery/ContentLoading.tsx
